Organize remaining config
