+++ src/lib/util/log.rs
@@ -2,22 +2,6 @@ use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 
 #[derive(Serialize, Deserialize, Clone)]
-#[serde(default)]
-pub struct LogConfig {
-    pub level: LogLevel,
-    pub format: String,
-}
-
-impl Default for LogConfig {
-    fn default() -> Self {
-        Self {
-            level: LogLevel::Warn,
-            format: r#"%a "%r" %s %b "%{Referer}i" "%{User-Agent}i" %T"#.to_string(),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone)]
 #[serde(rename_all = "lowercase")]
 pub enum LogLevel {
     Off,
