+++ src/lib/util/mod.rs
@@ -1,3 +1,4 @@
 pub mod command;
 pub mod config;
+pub mod log;
 pub mod shell;
